backend/scheduler-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
celery==5.4.0
redis==5.2.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
httpx==0.27.0
structlog==24.4.0
python-json-logger==3.2.1
colorama==0.4.6

# Monitoring
prometheus-client==0.21.0

# Message Queue
aio-pika==9.4.3

# Service Discovery
python-consul==1.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.27.0
